refactor: move to public package to be usable outside
