+++ pkg/instrumentation/get-target-id.go
@@ -1,4 +1,4 @@
-package utils
+package instrumentation
 
 import (
 	"crypto/sha256"
