+++ pkg/instrumentation/get-target-id_test.go
@@ -1,4 +1,4 @@
-package utils
+package instrumentation
 
 import (
 	"bufio"
